requirements.txt
openai>=1.0.0
elevenlabs
spotipy
RPi.GPIO
requests
python-dotenv
